Its able to load vocabulary and dotCode to memory
Now it can load both vocabularyCount.json and dotCode.json
to memory correctly. And the data is store in list type which is
called dataSet.

vocCountArrangeSamePerson will be renamed into
vocCountArrangeEachPerson in the next version. And this func
is used to arrange the data in vocabularyCount in each person.
just query person id then you can get the vocabulary he answered.