Finish difficultyDetect func and modify crawl.py
Func difficultyDetect in main.py is a function that can value
how hard a word is. The equation used in the func is copy from
the essay that NCKU ES presented. The equation may not reallly reliable.
However, we just used it as an init value for training.

In addition, in order to read data that been saved easily, the way to
save data into txt are all changed. They dont just save with f = open()
now. Now, I save them with pickle which allows me to read it really easy.

I remove the function call main() in crawl.py. Since that may be the
reason why main() run two times if I call the func in main.py